Add files via upload
Title: IPL Analysis Dashboard (2008-2020): Unveiling Insights into the Premier Cricketing Extravaganza.

Welcome to the IPL Analysis Dashboard (2008-2020), a comprehensive and dynamic tool designed to unravel the nuances and trends of the Indian Premier League (IPL) over the years. Immerse yourself in the world of cricket as you explore this powerful dashboard that brings together a wealth of data and analytics from the IPL's inaugural season in 2008 to the thrilling battles of 2020.

The IPL Analysis Dashboard offers a treasure trove of information and visualizations that cater to cricket enthusiasts, statisticians, team strategists, and fans alike. It provides an unprecedented level of detail and context, enabling you to understand the evolution of the IPL, track team and player performances, and uncover valuable insights that drive strategic decision-making.